update config and rename dummy repo files for create readme tests
