+++ ui-tests-starter/tst-243+/software/aws/toolkits/jetbrains/uitests/docTests/updateReadmeTests/UpdateReadmeLatestChanges/UpdateReadmeLatestChangesTest.kt
@@ -84,7 +84,7 @@ class UpdateReadmeLatestChangesTest {
                 val readme = File(readmePath.toUri())
                 assertTrue(readme.exists())
                 println("result: ${readme.readText()}")
-                assertTrue(readme.readText().contains("tancode", ignoreCase = true))
+                assertTrue(readme.readText().contains("sample/qdoc", ignoreCase = true))
                 assertTrue(readme.readText().contains("healthcontroller.java", ignoreCase = true))
                 println("assertions done")
             }
