+++ ui-tests-starter/tstData/qdoc/updateFlow/src/com/sample/qdoc/App.java
@@ -1,4 +1,4 @@
-package com.zetcode.tancode;
+package com.sample.qdoc;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
