+++ ui-tests-starter/tstData/qdoc/updateFlow/src/com/sample/qdoc/controller/HealthController.java
@@ -1,7 +1,7 @@
 // Copyright 2025 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.zetcode.tancode.controller;
+package com.sample.qdoc.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
