+++ ui-tests-starter/tstData/qdoc/updateFlow/src/com/sample/qdoc/model/DataItem.java
@@ -1,4 +1,4 @@
-package com.zetcode.tancode.model;
+package com.sample.qdoc.model;
 
 public class DataItem {
     private String id;
